~ UI
+ isKeyPressed()
+ RectangleF
~ PointF float -> double

+++ common/src/main/java/net/kroia/modutilities/gui/geometry/PointF.java
@@ -1,17 +1,17 @@
 package net.kroia.modutilities.gui.geometry;
 
 public class PointF {
-    public float x;
-    public float y;
-    public PointF(float x, float y) {
+    public double x;
+    public double y;
+    public PointF(double x, double y) {
         this.x = x;
         this.y = y;
     }
 
     public PointF getRotated(float angle) {
         double rad = Math.toRadians(angle);
-        float x = (float) (this.x * Math.cos(rad) - this.y * Math.sin(rad));
-        float y = (float) (this.x * Math.sin(rad) + this.y * Math.cos(rad));
+        double x = (this.x * Math.cos(rad) - this.y * Math.sin(rad));
+        double y = (this.x * Math.sin(rad) + this.y * Math.cos(rad));
         return new PointF(x, y);
     }
 }
